+++ typescript/cli/src/deploy/warp.ts
@@ -567,7 +567,8 @@ async function updateExistingWarpRoute(
 ) {
   logBlue('Updating deployed Warp Routes');
   const { multiProvider, registry } = params.context;
-  const addresses = await registry.getAddresses();
+  const registryAddresses =
+    (await registry.getAddresses()) as ChainMap<ProxyFactoryFactoriesAddresses>;
   const contractVerifier = new ContractVerifier(
     multiProvider,
     apiKeys,
@@ -586,15 +587,13 @@ async function updateExistingWarpRoute(
             `Missing artifacts for ${chain}. Probably new deployment. Skipping update...`,
           );
 
-        config.ismFactoryAddresses = addresses[
-          chain
-        ] as ProxyFactoryFactoriesAddresses;
         const evmERC20WarpModule = new EvmERC20WarpModule(
           multiProvider,
           {
             config,
             chain,
             addresses: {
+              ...registryAddresses[chain],
               deployedTokenRoute: deployedConfig.addressOrDenom!,
             },
           },
@@ -679,7 +678,9 @@ async function enrollRemoteRouters(
   deployedContractsMap: HyperlaneContractsMap<HypERC20Factories>,
 ): Promise<AnnotatedEV5Transaction[]> {
   logBlue(`Enrolling deployed routers with each other...`);
-  const { multiProvider } = params.context;
+  const { multiProvider, registry } = params.context;
+  const registryAddresses =
+    (await registry.getAddresses()) as ChainMap<ProxyFactoryFactoriesAddresses>;
   const deployedRoutersAddresses: ChainMap<Address> = objMap(
     deployedContractsMap,
     (_, contracts) => getRouter(contracts).address,
@@ -707,7 +708,10 @@ async function enrollRemoteRouters(
         const evmERC20WarpModule = new EvmERC20WarpModule(multiProvider, {
           config: mutatedWarpRouteConfig,
           chain,
-          addresses: { deployedTokenRoute: router.address },
+          addresses: {
+            ...registryAddresses[chain],
+            deployedTokenRoute: router.address,
+          },
         });
 
         const otherChains = multiProvider
@@ -970,7 +974,7 @@ async function getWarpApplySubmitter({
   context: WriteCommandContext;
   strategyUrl?: string;
 }): Promise<TxSubmitterBuilder<ProtocolType>> {
-  const { chainMetadata, multiProvider } = context;
+  const { multiProvider } = context;
 
   const submissionStrategy: SubmissionStrategy = strategyUrl
     ? readChainSubmissionStrategy(strategyUrl)[chain]
@@ -981,8 +985,7 @@ async function getWarpApplySubmitter({
         },
       };
 
-  const protocol = chainMetadata[chain].protocol;
-  return getSubmitterBuilder<typeof protocol>({
+  return getSubmitterBuilder<ProtocolType>({
     submissionStrategy,
     multiProvider,
   });