+++ typescript/infra/config/environments/mainnet3/agent.ts
@@ -12,6 +12,7 @@ import {
 import {
   MetricAppContext,
   routerMatchingList,
+  senderMatchingList,
   warpRouteMatchingList,
 } from '../../../src/config/agent/relayer.js';
 import { ALL_KEY_ROLES, Role } from '../../../src/roles.js';
@@ -20,14 +21,18 @@ import { getDomainId } from '../../registry.js';
 
 import { environment } from './chains.js';
 import { helloWorld } from './helloworld.js';
+import aaveSenderAddresses from './misc-artifacts/aave-sender-addresses.json';
+import everclearSenderAddresses from './misc-artifacts/everclear-sender-addresses.json';
+import merklyEthAddresses from './misc-artifacts/merkly-eth-addresses.json';
+import merklyNftAddresses from './misc-artifacts/merkly-eth-addresses.json';
+import merklyErc20Addresses from './misc-artifacts/merkly-eth-addresses.json';
+import veloMessageModuleAddresses from './misc-artifacts/velo-message-module-addresses.json';
+import veloTokenBridgeAddresses from './misc-artifacts/velo-token-bridge-addresses.json';
 import {
   mainnet3SupportedChainNames,
   supportedChainNames,
 } from './supportedChainNames.js';
 import { validatorChainConfig } from './validators.js';
-import merklyEthAddresses from './warp/artifacts/merkly-eth-addresses.json';
-import merklyNftAddresses from './warp/artifacts/merkly-eth-addresses.json';
-import merklyErc20Addresses from './warp/artifacts/merkly-eth-addresses.json';
 import { WarpRouteIds } from './warp/warpIds.js';
 
 // const releaseCandidateHelloworldMatchingList = routerMatchingList(
@@ -60,7 +65,9 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     bitlayer: true,
     blast: true,
     bob: true,
+    boba: true,
     bsc: true,
+    bsquared: true,
     celo: true,
     cheesechain: true,
     chilizmainnet: true,
@@ -68,6 +75,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     cyber: true,
     degenchain: true,
     dogechain: true,
+    duckchain: true,
     eclipsemainnet: true,
     endurance: true,
     ethereum: true,
@@ -120,9 +128,12 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     snaxchain: true,
     solanamainnet: true,
     stride: false,
+    superseed: true,
     superpositionmainnet: true,
     taiko: true,
     tangle: true,
+    unichain: true,
+    vana: true,
     viction: true,
     worldchain: true,
     xai: true,
@@ -148,7 +159,9 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     bitlayer: true,
     blast: true,
     bob: true,
+    boba: true,
     bsc: true,
+    bsquared: true,
     celo: true,
     cheesechain: true,
     chilizmainnet: true,
@@ -156,6 +169,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     cyber: true,
     degenchain: true,
     dogechain: true,
+    duckchain: true,
     eclipsemainnet: true,
     endurance: true,
     ethereum: true,
@@ -209,9 +223,12 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     snaxchain: true,
     solanamainnet: true,
     stride: true,
+    superseed: true,
     superpositionmainnet: true,
     taiko: true,
     tangle: true,
+    unichain: true,
+    vana: true,
     viction: true,
     worldchain: true,
     xai: true,
@@ -237,7 +254,9 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     bitlayer: true,
     blast: true,
     bob: true,
+    boba: true,
     bsc: true,
+    bsquared: true,
     celo: true,
     cheesechain: true,
     chilizmainnet: true,
@@ -245,6 +264,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     cyber: true,
     degenchain: true,
     dogechain: true,
+    duckchain: true,
     // Cannot scrape Sealevel chains
     eclipsemainnet: false,
     endurance: true,
@@ -299,9 +319,12 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     // Cannot scrape Sealevel chains
     solanamainnet: false,
     stride: true,
+    superseed: true,
     superpositionmainnet: true,
     taiko: true,
     tangle: true,
+    unichain: true,
+    vana: true,
     // Has RPC non-compliance that breaks scraping.
     viction: false,
     worldchain: true,
@@ -375,6 +398,28 @@ const metricAppContextsGetter = (): MetricAppContext[] => {
       name: 'merkly_nft',
       matchingList: routerMatchingList(merklyNftAddresses),
     },
+    {
+      name: 'velo_message_module',
+      matchingList: routerMatchingList(veloMessageModuleAddresses),
+    },
+    {
+      name: 'velo_token_bridge',
+      matchingList: routerMatchingList(veloTokenBridgeAddresses),
+    },
+    {
+      // https://github.com/bgd-labs/aave-delivery-infrastructure?tab=readme-ov-file#deployed-addresses
+      // We match on senders because the sender is always the same and
+      // well documented, while the recipient may be switched out and is
+      // more poorly documented.
+      name: 'aave',
+      matchingList: senderMatchingList(aaveSenderAddresses),
+    },
+    {
+      // https://docs.everclear.org/resources/contracts/mainnet
+      // Messages between HubGateway (Everclear hub) <> EverclearSpoke (all other spoke chains)
+      name: 'everclear_gateway',
+      matchingList: senderMatchingList(everclearSenderAddresses),
+    },
   ];
 };
 
@@ -409,7 +454,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '75d62ae-20241107-060707',
+      tag: 'da277d9-20241128-121942',
     },
     gasPaymentEnforcement: gasPaymentEnforcement,
     metricAppContextsGetter,
@@ -418,7 +463,7 @@ const hyperlane: RootAgentConfig = {
   validators: {
     docker: {
       repo,
-      tag: '75d62ae-20241107-060707',
+      tag: 'd834d81-20241125-135658',
     },
     rpcConsensusType: RpcConsensusType.Quorum,
     chains: validatorChainConfig(Contexts.Hyperlane),
@@ -428,7 +473,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '75d62ae-20241107-060707',
+      tag: 'd834d81-20241125-135658',
     },
     resources: scraperResources,
   },
@@ -443,7 +488,7 @@ const releaseCandidate: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '75d62ae-20241107-060707',
+      tag: '25a927d-20241114-171323',
     },
     // We're temporarily (ab)using the RC relayer as a way to increase
     // message throughput.
@@ -476,7 +521,7 @@ const neutron: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '75d62ae-20241107-060707',
+      tag: '25a927d-20241114-171323',
     },
     gasPaymentEnforcement,
     metricAppContextsGetter,