+++ typescript/infra/config/environments/mainnet3/warp/configGetters/getAncient8EthereumUSDCWarpConfig.ts
@@ -2,17 +2,21 @@ import { ethers } from 'ethers';
 
 import {
   ChainMap,
-  RouterConfig,
+  OwnableConfig,
   TokenRouterConfig,
   TokenType,
   buildAggregationIsmConfigs,
   defaultMultisigConfigs,
 } from '@hyperlane-xyz/sdk';
 
-import { tokens } from '../../../../../src/config/warp.js';
+import {
+  RouterConfigWithoutOwner,
+  tokens,
+} from '../../../../../src/config/warp.js';
 
 export const getAncient8EthereumUSDCWarpConfig = async (
-  routerConfig: ChainMap<RouterConfig>,
+  routerConfig: ChainMap<RouterConfigWithoutOwner>,
+  abacusWorksEnvOwnerConfig: ChainMap<OwnableConfig>,
 ): Promise<ChainMap<TokenRouterConfig>> => {
   const ismConfig = buildAggregationIsmConfigs(
     'ethereum',
@@ -22,6 +26,7 @@ export const getAncient8EthereumUSDCWarpConfig = async (
 
   const ethereum: TokenRouterConfig = {
     ...routerConfig.ethereum,
+    ...abacusWorksEnvOwnerConfig.ethereum,
     type: TokenType.collateral,
     token: tokens.ethereum.USDC,
     interchainSecurityModule: ismConfig,
@@ -31,12 +36,9 @@ export const getAncient8EthereumUSDCWarpConfig = async (
     hook: '0x19b2cF952b70b217c90FC408714Fbc1acD29A6A8',
   };
 
-  // @ts-ignore - The types as they stand require a synthetic to specify
-  // TokenMetadata, but in practice these are actually inferred from a
-  // collateral config. To avoid needing to specify the TokenMetadata, just
-  // ts-ignore for synthetic tokens.
   const ancient8: TokenRouterConfig = {
     ...routerConfig.ancient8,
+    ...abacusWorksEnvOwnerConfig.ancient8,
     type: TokenType.synthetic,
     // Uses the default ISM
     interchainSecurityModule: ethers.constants.AddressZero,