+++ typescript/infra/config/environments/mainnet3/warp/configGetters/getInevmInjectiveINJWarpConfig.ts
@@ -1,23 +1,28 @@
 import {
   ChainMap,
-  RouterConfig,
+  OwnableConfig,
   TokenRouterConfig,
   TokenType,
 } from '@hyperlane-xyz/sdk';
 
+import { RouterConfigWithoutOwner } from '../../../../../src/config/warp.js';
+
 export const getInevmInjectiveINJWarpConfig = async (
-  routerConfig: ChainMap<RouterConfig>,
+  routerConfig: ChainMap<RouterConfigWithoutOwner>,
+  abacusWorksEnvOwnerConfig: ChainMap<OwnableConfig>,
 ): Promise<ChainMap<TokenRouterConfig>> => {
   const injectiveRouter = 'inj1mv9tjvkaw7x8w8y9vds8pkfq46g2vcfkjehc6k';
 
-  // @ts-ignore - foreignDeployment configs don't conform to the TokenRouterConfig
   const injective: TokenRouterConfig = {
+    ...routerConfig.injective,
+    ...abacusWorksEnvOwnerConfig.injective,
     type: TokenType.native,
     foreignDeployment: injectiveRouter,
   };
 
   const inevm: TokenRouterConfig = {
     ...routerConfig.inevm,
+    ...abacusWorksEnvOwnerConfig.inevm,
     type: TokenType.native,
   };