+++ typescript/infra/scripts/agent-utils.ts
@@ -1,11 +1,8 @@
+import { checkbox, select } from '@inquirer/prompts';
 import path, { join } from 'path';
 import yargs, { Argv } from 'yargs';
 
-import {
-  ChainAddresses,
-  IRegistry,
-  warpConfigToWarpAddresses,
-} from '@hyperlane-xyz/registry';
+import { ChainAddresses, IRegistry } from '@hyperlane-xyz/registry';
 import {
   ChainMap,
   ChainMetadata,
@@ -27,6 +24,7 @@ import {
 
 import { Contexts } from '../config/contexts.js';
 import { agents } from '../config/environments/agents.js';
+import { WarpRouteIds } from '../config/environments/mainnet3/warp/warpIds.js';
 import { validatorBaseConfigsFn } from '../config/environments/utils.js';
 import {
   getChain,
@@ -157,20 +155,26 @@ export function withChain<T>(args: Argv<T>) {
     .alias('c', 'chain');
 }
 
-export function withChains<T>(args: Argv<T>) {
+export function withChains<T>(args: Argv<T>, chainOptions?: ChainName[]) {
   return (
     args
       .describe('chains', 'Set of chains to perform actions on.')
       .array('chains')
-      .choices('chains', getChains())
+      .choices(
+        'chains',
+        !chainOptions || chainOptions.length === 0 ? getChains() : chainOptions,
+      )
       // Ensure chains are unique
       .coerce('chains', (chains: string[]) => Array.from(new Set(chains)))
       .alias('c', 'chains')
   );
 }
 
-export function withChainsRequired<T>(args: Argv<T>) {
-  return withChains(args).demandOption('chains');
+export function withChainsRequired<T>(
+  args: Argv<T>,
+  chainOptions?: ChainName[],
+) {
+  return withChains(args, chainOptions).demandOption('chains');
 }
 
 export function withWarpRouteId<T>(args: Argv<T>) {
@@ -191,9 +195,8 @@ export function withProtocol<T>(args: Argv<T>) {
 
 export function withAgentRole<T>(args: Argv<T>) {
   return args
-    .describe('role', 'agent roles')
-    .array('role')
-    .coerce('role', (role: string[]): Role[] => role.map(assertRole))
+    .describe('role', 'agent role')
+    .coerce('role', (role: string): Role => assertRole(role))
     .demandOption('role')
     .alias('r', 'role');
 }
@@ -206,9 +209,14 @@ export function withAgentRoles<T>(args: Argv<T>) {
       .coerce('roles', (role: string[]): Role[] => role.map(assertRole))
       .choices('roles', Object.values(Role))
       // Ensure roles are unique
-      .coerce('roles', (roles: string[]) => Array.from(new Set(roles)))
+      .coerce('roles', (roles: Role[]) => Array.from(new Set(roles)))
       .alias('r', 'roles')
+      .alias('role', 'roles')
   );
+}
+
+export function withAgentRolesRequired<T>(args: Argv<T>) {
+  return withAgentRoles(args).demandOption('roles');
 }
 
 export function withKeyRoleAndChain<T>(args: Argv<T>) {
@@ -262,6 +270,49 @@ export function withRpcUrls<T>(args: Argv<T>) {
     .string('rpcUrls')
     .demandOption('rpcUrls')
     .alias('r', 'rpcUrls');
+}
+
+export function withTxHashes<T>(args: Argv<T>) {
+  return args
+    .describe('txHashes', 'transaction hash')
+    .string('txHashes')
+    .array('txHashes')
+    .demandOption('txHashes')
+    .alias('t', 'txHashes');
+}
+
+// Interactively gets a single warp route ID
+export async function getWarpRouteIdInteractive() {
+  const choices = Object.values(WarpRouteIds).map((id) => ({
+    value: id,
+  }));
+  return select({
+    message: 'Select Warp Route ID',
+    choices,
+    pageSize: 30,
+  });
+}
+
+// Interactively gets multiple warp route IDs
+export async function getWarpRouteIdsInteractive() {
+  const choices = Object.values(WarpRouteIds).map((id) => ({
+    value: id,
+  }));
+
+  let selection: WarpRouteIds[] = [];
+
+  while (!selection.length) {
+    selection = await checkbox({
+      message: 'Select Warp Route IDs',
+      choices,
+      pageSize: 30,
+    });
+    if (!selection.length) {
+      console.log('Please select at least one Warp Route ID');
+    }
+  }
+
+  return selection;
 }
 
 // not requiring to build coreConfig to get agentConfig