+++ typescript/infra/scripts/warp-routes/monitor/metrics.ts
@@ -3,7 +3,9 @@ import { Gauge, Registry } from 'prom-client';
 import { createWarpRouteConfigId } from '@hyperlane-xyz/registry';
 import { ChainName, Token, TokenStandard, WarpCore } from '@hyperlane-xyz/sdk';
 
-import { WarpRouteBalance, XERC20Limit } from './types.js';
+import { getWalletBalanceGauge } from '../../../src/utils/metrics.js';
+
+import { NativeWalletBalance, WarpRouteBalance, XERC20Limit } from './types.js';
 import { logger } from './utils.js';
 
 export const metricsRegister = new Registry();
@@ -44,6 +46,8 @@ const warpRouteCollateralValue = new Gauge({
   labelNames: warpRouteMetricLabels,
 });
 
+const walletBalanceGauge = getWalletBalanceGauge(metricsRegister);
+
 const xERC20LimitsGauge = new Gauge({
   name: 'hyperlane_xerc20_limits',
   help: 'Current minting and burning limits of xERC20 tokens',
@@ -74,18 +78,39 @@ export function updateTokenBalanceMetrics(
   };
 
   warpRouteTokenBalance.labels(metrics).set(balanceInfo.balance);
-  logger.info('Wallet balance updated for token', {
-    labels: metrics,
-    balance: balanceInfo.balance,
-  });
+  logger.info(
+    {
+      labels: metrics,
+      balance: balanceInfo.balance,
+    },
+    'Wallet balance updated for token',
+  );
 
   if (balanceInfo.valueUSD) {
     warpRouteCollateralValue.labels(metrics).set(balanceInfo.valueUSD);
-    logger.info('Wallet balance updated for token', {
-      labels: metrics,
-      valueUSD: balanceInfo.valueUSD,
-    });
+    logger.info(
+      {
+        labels: metrics,
+        valueUSD: balanceInfo.valueUSD,
+      },
+      'Wallet value updated for token',
+    );
   }
+}
+
+export function updateNativeWalletBalanceMetrics(balance: NativeWalletBalance) {
+  walletBalanceGauge
+    .labels({
+      chain: balance.chain,
+      wallet_address: balance.walletAddress,
+      wallet_name: balance.walletName,
+      token_symbol: 'Native',
+      token_name: 'Native',
+    })
+    .set(balance.balance);
+  logger.info('Native wallet balance updated', {
+    balanceInfo: balance,
+  });
 }
 
 export function updateXERC20LimitsMetrics(token: Token, limits: XERC20Limit) {