+++ typescript/utils/src/index.ts
@@ -26,6 +26,7 @@ export {
   isValidAddressCosmos,
   isValidAddressEvm,
   isValidAddressSealevel,
+  isPrivateKeyEvm,
   isValidTransactionHash,
   isValidTransactionHashCosmos,
   isValidTransactionHashEvm,
@@ -108,6 +109,7 @@ export {
   arrayToObject,
   deepCopy,
   deepEquals,
+  deepFind,
   diffObjMerge,
   invertKeysAndValues,
   isObjEmpty,
@@ -139,6 +141,7 @@ export {
   streamToString,
   toHexString,
   toTitleCase,
+  toUpperCamelCase,
   trimToLength,
 } from './strings.js';
 export { isNullish, isNumeric } from './typeof.js';
@@ -168,9 +171,10 @@ export {
   S3CheckpointWithId,
   SignatureLike,
   TokenCaip19Id,
+  ValidatorMetadata,
   WithAddress,
 } from './types.js';
-export { isHttpsUrl, isUrl } from './url.js';
+export { isHttpsUrl, isRelativeUrl, isUrl } from './url.js';
 export { assert } from './validation.js';
 export { BaseValidator, ValidatorConfig } from './validator.js';
 export { tryParseJsonOrYaml } from './yaml.js';