+++ typescript/utils/src/types.ts
@@ -118,3 +118,7 @@ export type ParsedLegacyMultisigIsmMetadata = {
 export type Annotated<T> = T & {
   annotation?: string;
 };
+
+export type ValidatorMetadata = {
+  git_sha: string;
+};